Finished EX_03 and Corrected EX_02

+++ APLesson_03/EX_02.java
@@ -7,9 +7,9 @@ public class EX_02
 		Scanner keyboard = new Scanner(System.in);
 		
 		System.out.println("What is your height in inches?");
-		int height = keyboard.nextInt();
+		float height = keyboard.nextInt();
 		System.out.println("What is your weight in pounds?");
-		int weight = keyboard.nextInt();
+		float weight = keyboard.nextInt();
 		
 		float BMI = (weight*703)/(height*height);
